pyproject.toml
[project]
name = "fraud-pipeline-patrol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-core>=1.9.4",
    "dbt-coverage>=0.3.9",
    "dbt-duckdb>=1.9.3",
    "dbterd",
    "duckdb>=1.2.2",
    "fastparquet>=2024.11.0",
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "minio>=7.2.15",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
    "pre-commit>=3.0.0",
    "types-requests>=2.0.0",
]
ci = [
    "apache-airflow==2.8.1",
]

[tool.uv.sources]
dbterd = { git = "https://github.com/intellishore/dbt-erdiagram-generator", subdirectory = "src" }

[tool.uv.workspace]
members = [
    "dbt/fraud_detection",
]

[tool.setuptools.packages.find]
include = ["fraud_detection"]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["fraud_pipeline_patrol", "helpers", "scripts"]

# pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true
